feat: separate word validation
Separates word validation, the green and yellow letter logic, into a helper file.
